feat: accept nested properties in mp.get()
